+++ problem/assignment.go
@@ -1,8 +1,13 @@
 package problem
 
 import (
+	"fmt"
+	"slices"
+
 	"github.com/roidaradal/fn/comb"
 	"github.com/roidaradal/fn/list"
+	"github.com/roidaradal/fn/number"
+	"github.com/roidaradal/fn/str"
 	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
@@ -16,10 +21,14 @@ func NewAssignment(variant string, n int) *discrete.Problem {
 		return assignment(name)
 	case "bottleneck":
 		return bottleneckAssignment(name)
+	case "general":
+		return generalizedAssignment(name)
 	case "quadratic":
 		return quadraticAssignment(name)
 	case "quadratic_bottleneck":
 		return quadraticBottleneckAssignment(name)
+	case "weapon":
+		return weaponTargetAssignment(name)
 	default:
 		return nil
 	}
@@ -27,7 +36,7 @@ func NewAssignment(variant string, n int) *discrete.Problem {
 
 // Common steps for creating Assignment problem
 func newAssignmentProblem(name string) (*discrete.Problem, *data.AssignmentCfg) {
-	cfg := data.NewAssignment(name)
+	cfg := data.NewAssignment(name, true)
 	if cfg == nil {
 		return nil, nil
 	}
@@ -104,6 +113,54 @@ func bottleneckAssignment(name string) *discrete.Problem {
 	return p
 }
 
+// Generalized Assignment
+func generalizedAssignment(name string) *discrete.Problem {
+	cfg := data.NewAssignment(name, false)
+	if cfg == nil {
+		return nil
+	}
+
+	p := discrete.NewProblem(name)
+	p.Type = discrete.Assignment
+
+	p.Variables = discrete.Variables(cfg.Tasks)
+	p.AddVariableDomains(discrete.Domain(cfg.Workers))
+
+	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
+		// Compute total cost of assigning tasks to each worker
+		total := make(map[int]float64)
+		for task, worker := range solution.Map {
+			total[worker] += cfg.Cost[worker][task]
+		}
+		// Check that worker totals don't exceed their capacity
+		for worker, limit := range cfg.Capacity {
+			if total[worker] > limit {
+				return false
+			}
+		}
+		return true
+	})
+
+	p.Goal = discrete.Maximize
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		// Total value of assigning tasks to workers
+		var totalValue discrete.Score = 0
+		for task, worker := range solution.Map {
+			totalValue += cfg.Value[worker][task]
+		}
+		return totalValue
+	}
+
+	p.SolutionStringFn = func(solution *discrete.Solution) string {
+		output := list.Map(p.Variables, func(task discrete.Variable) string {
+			worker := solution.Map[task]
+			return fmt.Sprintf("t%s = w%s", cfg.Tasks[task], cfg.Workers[worker])
+		})
+		return str.WrapBraces(output)
+	}
+	return p
+}
+
 // Common steps for creating Quadratic Assignment problem
 func newQuadraticAssignmentProblem(name string) (*discrete.Problem, *data.QuadraticAssignment) {
 	cfg := data.NewQuadraticAssignment(name)
@@ -166,5 +223,67 @@ func quadraticBottleneckAssignment(name string) *discrete.Problem {
 		}
 		return maxCost
 	}
+	return p
+}
+
+// Weapon Target Assignment
+func weaponTargetAssignment(name string) *discrete.Problem {
+	cfg := data.NewWeapons(name)
+	if cfg == nil {
+		return nil
+	}
+	numWeapons, numTargets := len(cfg.Weapons), len(cfg.Targets)
+
+	p := discrete.NewProblem(name)
+	p.Type = discrete.Assignment
+
+	// Expand weapons and count into weapons list (uses weapon index)
+	weapons := make([]int, 0)
+	for i := range numWeapons {
+		weapons = append(weapons, slices.Repeat([]int{i}, cfg.Count[i])...)
+	}
+
+	p.Variables = discrete.Variables(weapons)
+	p.AddVariableDomains(discrete.Domain(cfg.Targets))
+
+	p.Goal = discrete.Minimize
+	p.ObjectiveFn = func(solution *discrete.Solution) discrete.Score {
+		// Compute survival rate of each target, with weapons assigned to attack it
+		survival := list.Copy(cfg.Value)
+		for w, target := range solution.Map {
+			weapon := weapons[w]
+			survival[target] *= 1 - cfg.Chance[weapon][target]
+			// survival = 1 - weaponOnTargetEffectiveness
+		}
+		total := fmt.Sprintf("%.4f", list.Sum(survival))
+		return number.ParseFloat(total)
+	}
+
+	weaponTargets := func(solution *discrete.Solution) string {
+		// Group count of weapon => target assignments
+		matrix := make([][]int, numWeapons)
+		for i := range numWeapons {
+			matrix[i] = make([]int, numTargets)
+		}
+		for w, target := range solution.Map {
+			weapon := weapons[w]
+			matrix[weapon][target] += 1
+		}
+		output := make([]string, 0)
+		for i, weapon := range cfg.Weapons {
+			for j, target := range cfg.Targets {
+				if matrix[i][j] == 0 {
+					continue // skip empty count
+				}
+				line := fmt.Sprintf("%d*%s = %s", matrix[i][j], weapon, target)
+				output = append(output, line)
+			}
+		}
+		return str.WrapBraces(output)
+	}
+
+	p.SolutionStringFn = weaponTargets
+	p.SolutionCoreFn = weaponTargets
+
 	return p
 }